webv2/requirements.txt
adb-shell==0.4.4
celery==5.3.6
fastapi==0.109.2
libvirt-python==10.0.0
orjson==3.9.15
pydantic==2.6.0
redis==5.0.1
requests==2.31.0
starlette==0.36.3
uvicorn==0.27.0.post1
uvloop==0.19.0
